設定ファイルの入出力をリファクタ
